% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{export_data}
\alias{export_data}
\title{A function to export data frames to separate Google sheets}
\usage{
export_data(data_list, sheet_id = "create", names = NULL, sheet_num = 1)
}
\description{
This function exports a list of data frames into their own Google sheets. It can
either create new Google sheets or write into pre-existing ones.
}
