% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pokeacademy_data}
\alias{pokeacademy_data}
\title{Pokemon Academy Data}
\format{
\subsection{\code{pokeacademy_data}}{

A data frame with 599 rows and 9 columns
\describe{
\item{Last_Name}{The last name of the student}
\item{First_Name}{The first name of the student}
\item{Student_Type}{The type of student, either an entering first year or entering transfer.}
\item{Hall}{The residential hall that the student lives in}
\item{Room}{The student's room number}
\item{Room_Type}{The type of room the student lives in: Single, Double, Triple, or Quadruple}
\item{Citizenship_DESC}{Description of the student's citizenship: Domestic or International}
\item{DOB}{The student's date of birth}
\item{Email}{The student's school email address}
}
}
}
\source{
Generated by package creators
}
\usage{
pokeacademy_data
}
\description{
A completely made-up data set of information about students entering the pokemon academy
}
\keyword{datasets}
