% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_sheet.R
\name{split_sheet}
\alias{split_sheet}
\title{Split a Data Frame by Specified Variables}
\usage{
split_sheet(sheet, ..., keep = FALSE)
}
\arguments{
\item{sheet}{A data frame (generally read in from a Google sheet) to be split upon one or more
specified variables.}

\item{...}{The groups to be split on. Generally, column names in the \code{data_frame$variable}
format. When more than one group is provided, the data frame is split on their interaction
(i.e. split into groups resulting from each combination of values of the groups/variables being
split on).}

\item{keep}{A logical vector of length one indicating whether to keep all data frames (if TRUE) or
discard data frames for groups where there are no observations (if FALSE), particularly if more than one
variable is specified for splitting.}
}
\value{
A list of data frames that have been split on one or more specified variables. The
components are named after the categories they have been split on (as seen in the
\code{\link[base:split]{base::split}} method.)
}
\description{
This function splits a data frame into a list of smaller data frames based on one or more
specified variables.
}
\examples{
\dontrun{
  df <- data.frame (A = c("a", "a", "b", "b"),
                    B = c(1, 2, 2, 3),
                    C = c(TRUE, TRUE, TRUE, FALSE))
  ## Don't keep data frames with no observations (ex: group formed by df$A == "a" and df$B == 3)
  data_list <- split_sheet(df, df$A, df$B)

  ## Keep all data frames regardless of if they are empty.
  data_list_2 <- split_sheet(df, df$A, df$B, keep = TRUE)
}

}
