% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{export_data}
\alias{export_data}
\title{A function to export data frames to separate Google sheets}
\usage{
export_data(data_list, sheet_id = "create", names = NULL, sheet_num = 1)
}
\arguments{
\item{data_list}{A list of data frames to be exported to separate Google sheets.}

\item{sheet_id}{A vector of existing Google sheet ids, urls to extract sheet ids from or
\code{googlesheets4} ss objects. Alternatively, "create" in order to write new Google sheets.}

\item{names}{A vector with the same length as data_list which exported Google sheets will be
named with. Can be \code{NULL}, in which case the names will be "Untitled_Sheet_i" where \code{i} is the
numeric index in the list of the data frame being exported.}

\item{sheet_num}{A numeric value representing the worksheet within the larger Google sheet that
will be written to.}
}
\value{
A character vector of sheet IDs that the function has exported data to. This will be the
input if the data was exported to an existing Google sheet and the new sheet IDs if the data
was exported to new Google sheets.
}
\description{
This function exports a list of data frames into their own Google sheets. It can
either create new Google sheets or write into pre-existing ones.
}
\examples{
\dontrun{
    df1 <- data.frame(A = c("a", "a", "b", "b"),
                  B = c(1, 2, 2, 3),
                  C = c(TRUE, TRUE, TRUE, FALSE))
    df2 <- data.frame(A = c("a", "a", "b", "b"),
                  B = c(1, 2, 2, 3),
                  C = c(TRUE, TRUE, TRUE, FALSE))
    df_list <- list(df1, df2)


    # Export data to existing Google sheets
    export_data(df_list, sheet_id = c("export data demo 1", "export data demo 2"))

    # Export data to new Google sheets
    export_data(df_list, names = c("Demo_Sheet_1", "Demo_Sheet_2"))
}

}
